perf(fast_time): use u64 nanosecond clock arithmetic on Linux (#219)
## Summary

Switch the Linux `fast_time` clock source from `u128` to `u64`
nanosecond
arithmetic. `u64` nanoseconds spans ~584 years, far beyond the range of
`CLOCK_MONOTONIC_COARSE`, so the wider type bought us nothing while
costing
extra instructions on every `Clock::now()` call.

The change:

* `Bindings::clock_gettime_nanos` now returns `u64` (abstraction,
facade, real).
* `TimeSourceImpl` stores `platform_epoch` / `cache_key` as `u64`.
* Removes the `u64::try_from(...).expect(...)` narrowing in `now()` and
the
  associated `u128` multiply/add in the real binding.

Windows already used `u64` (`get_tick_count_64`), so no mirrored change
is
needed — this only brings Linux in line.

## Measurement

Callgrind (`just package=fast_time bench-cg`), deterministic instruction
counts:

| scenario    | before | after | delta    |
|-------------|-------:|------:|----------|
| `clock_now` |    108 |    88 | **−18.5%** |

`clock_now` uses a fresh `Clock`, so it always takes the cache-miss
arithmetic path — a deterministic, reproducible measurement.

(The `instant_elapsed` scenario is timing-sensitive — its instruction
count
depends on whether `CLOCK_MONOTONIC_COARSE` ticks during the Valgrind
run, so
it flips between the cache-hit and cache-miss path between runs. Filed
separately; not a regression from this change, which can only reduce
work on
both paths.)

## Validation

`just package=fast_time validate-local` passes on Linux.

+++ packages/fast_time/src/pal/linux/bindings/abstractions.rs
@@ -6,7 +6,7 @@ use std::time::Instant;
 /// All PAL FFI calls must go through this trait, enabling them to be mocked.
 #[cfg_attr(test, mockall::automock)]
 pub(crate) trait Bindings: Debug + Send + Sync + 'static {
-    fn clock_gettime_nanos(&self) -> u128;
+    fn clock_gettime_nanos(&self) -> u64;
 
     // We also put this here because Rust does not (yet) support a proper clock abstraction,
     // so without this we have nothing to mock. This just provides a mock wrapper around `Instant`.
